requirements.txt
streamlit
sentence-transformers
chromadb
gpt4all
PyPDF2
pdfplumber
numpy
